Fixed tests in events/
