+++ lib/events/event_handler.js
@@ -3,9 +3,10 @@
 import fs from 'fs';
 import path from 'path';
 import getBranchName from "current-git-branch";
+import {DEFAULT_BRANCH} from "../constants";
 import { diff_match_patch } from "diff-match-patch";
-import { manageDiff, handleDirectoryRenameDiffs, handleDirectoryDeleteDiffs } from "./diff_utils";
-import { repoIsNotSynced, shouldIgnoreFile } from './utils';
+import { manageDiff, handleDirectoryDeleteDiffs } from "./diff_utils";
+import { handleNewFile, handleRename, repoIsNotSynced, shouldIgnoreFile } from './utils';
 import { generateSettings } from "../settings";
 
 
@@ -60,35 +61,8 @@ export function handleFileCreated(event) {
     if (!repoPath || repoIsNotSynced(repoPath) || shouldIgnoreFile(repoPath, relPath)) {
         return;
     }
-
-    // Do not continue if file does not exist
-    if (!fs.existsSync(filePath)) { return; }
-    // Do not continue if directoy is created
-    if (fs.lstatSync(filePath).isDirectory()) { return; }
-
-    const settings = generateSettings();
-    console.log(`FileCreated: ${filePath}`);
-    const branch = getBranchName({altPath: repoPath});
-    const destOriginals = `${settings.ORIGINALS_REPO}/${repoPath.slice(1)}/${branch}/${relPath}`;
-    const destOriginalsPathSplit = destOriginals.split("/");
-    const destOriginalsBasePath = destOriginalsPathSplit.slice(0, destOriginalsPathSplit.length - 1).join("/");
-    const destShadow = `${settings.SHADOW_REPO}/${repoPath.slice(1)}/${branch}/${relPath}`;
-    const destShadowPathSplit = destShadow.split("/");
-    const destShadowBasePath = destShadowPathSplit.slice(0, destShadowPathSplit.length - 1).join("/");
-    // Skip if file is already in .originals
-    if (fs.existsSync(destOriginals)) {
-        return;
-    }
-    // Add file in originals repo
-    fs.mkdirSync(destOriginalsBasePath, {recursive: true});
-    // File destination will be created or overwritten by default.
-    fs.copyFileSync(filePath, destOriginals);
-    // Add file in shadow repo
-    fs.mkdirSync(destShadowBasePath, {recursive: true});
-    // File destination will be created or overwritten by default.
-    fs.copyFileSync(filePath, destShadow);
-
-    manageDiff(repoPath, branch, relPath, "", true);
+    const branch = getBranchName({ altPath: repoPath }) || DEFAULT_BRANCH;
+    handleNewFile(repoPath, branch, filePath);
 }
 
 export function handleFileDeleted(event) {
@@ -119,7 +93,7 @@ export function handleFileDeleted(event) {
     console.log(`FileDeleted: ${itemPath}`);
 
     // Cache path
-    const destDeleted = path.join(settings.DELETED_REPO, `${repoPath}/${branch}/${relPath}`);;
+    const destDeleted = path.join(settings.DELETED_REPO, `${repoPath}/${branch}/${relPath}`);
     const destDeletedPathSplit = destDeleted.split("/");
     const destDeletedBasePath = destDeletedPathSplit.slice(0, destDeletedPathSplit.length-1).join("/");
 
@@ -145,30 +119,4 @@ export function handleFileRenamed(event) {
     }
     const branch = getBranchName({altPath: repoPath});
     handleRename(repoPath, branch, oldAbsPath, newAbsPath, fs.lstatSync(newAbsPath).isFile());
-}
-
-function handleRename(repoPath, branch, oldAbsPath, newAbsPath, isFile) {
-    const settings = generateSettings();
-    const oldRelPath = oldAbsPath.split(`${repoPath}/`)[1];
-    const newRelPath = newAbsPath.split(`${repoPath}/`)[1];
-    const oldShadowPath = `${settings.SHADOW_REPO}/${repoPath.slice(1)}/${branch}/${oldRelPath}`;
-    const newShadowPath = `${settings.SHADOW_REPO}/${repoPath.slice(1)}/${branch}/${newRelPath}`;
-    fs.renameSync(oldShadowPath, newShadowPath);
-
-    if (!isFile) {
-        console.log(`DirectoryRenamed: ${oldAbsPath} -> ${newAbsPath}`);
-        const diff = JSON.stringify({ old_path: oldAbsPath, new_path: newAbsPath });
-		handleDirectoryRenameDiffs(repoPath, branch, diff);
-        return;
-    }
-
-    console.log(`FileRenamed: ${oldAbsPath} -> ${newAbsPath}`);
-    const diff = JSON.stringify({
-        old_abs_path: oldAbsPath,
-        new_abs_path: newAbsPath,
-        old_rel_path: oldRelPath,
-        new_rel_path: newRelPath
-    });
-    // Add new diff in the buffer
-    manageDiff(repoPath, branch, newRelPath, diff, false, true);
 }
